Updates in SplunkUtils: 1. spelling correction - SplunkUtills to SplunkUtils (including in GTFS_find_changes.ipynb). 2. adding loop_over_splunk function for high volume queries. 3. adding planned start date column to Getting started with SplunkUitls.ipynb.
