+++ splunk/SplunkUtils/README.md
@@ -1,7 +1,7 @@
-SplunkUtills is an API for Splunk database, which can be used for querying SIRI and GTFS data into Pandas DataFrame.
+SplunkUtils is an API for Splunk database, which can be used for querying SIRI and GTFS data into Pandas DataFrame.
 
 It's requires installing splunk-sdk package.
 
 You must have Splunk credentials (host, port, username, password) to set a connection to Splunk (Note that the port is Splunk management port and not web port).
 
-To start go to **Getting started with SplunkUtills.ipynb.**  
+To start go to **Getting started with SplunkUtils.ipynb.**
